requirements.txt
openai>=1.0.0
python-dotenv
fastapi
uvicorn[standard]
pydantic>=2.0
